Base readable implem uses annotations
